% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_LSTM.R
\name{data_feature_func_LSTM}
\alias{data_feature_func_LSTM}
\title{Feature Extraction (LSTM)}
\usage{
data_feature_func_LSTM(data_raw, lookback_time, k, z)
}
\arguments{
\item{data_raw}{A list containing data for different cumulative time windows, which is the result returned by the function data_func().}

\item{lookback_time}{A vector representing the length of the cumulative time window.}

\item{k}{The number of neurons in the LSTM hidden layer.}

\item{z}{The number of neurons in the fully connected layer.}
}
\value{
A list containing the feature data frames extracted for each time window.
}
\description{
Feature Extraction (LSTM)
}
