% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{data_func}
\alias{data_func}
\title{Data Extraction Function}
\usage{
data_func(data, data0, pre_time, lookback_time)
}
\arguments{
\item{data}{Sepsis data, where records are continuous (in hours) and include the first hour of SIRS.}

\item{data0}{non-sepsis data, formatted similarly to param data.}

\item{pre_time}{The time period not considered in the model.}

\item{lookback_time}{A vector representing the length of the cumulative time window.}
}
\value{
Extracted data, where each element corresponds to the data for a specific time window.
}
\description{
Data Extraction Function
}
