% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_sep_data}
\alias{eg_sep_data}
\title{Sepsis Patient Dataset
\describe{
\item{n_var}{Number of variables: 12}
\item{n_hadm}{Number of unique hadm_id: 300}
}}
\source{
MIMIC-IV Database
}
\description{
Sepsis Patient Dataset
\describe{
\item{n_var}{Number of variables: 12}
\item{n_hadm}{Number of unique hadm_id: 300}
}
}
