% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.CTSM.R
\name{model2netcdf.CTSM}
\alias{model2netcdf.CTSM}
\title{Code to convert CTSM netcdf output into into CF standard}
\usage{
model2netcdf.CTSM(outdir)
}
\arguments{
\item{outdir}{Location of CTSM model output}
}
\description{
Code to convert CTSM netcdf output into into CF standard
}
\examples{
 
\dontrun{
example.output <- system.file("case.clm2.h0.2004-01-01-00000.nc",package="PEcAn.CTSM")
model2netcdf.CTSM(outdir="~/")
}

}
\author{
Michael Dietze, Shawn Serbin
}
