% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train_test.R
\name{stacking_func0}
\alias{stacking_func0}
\title{Cum-Time Stacking Model Training and Prediction}
\usage{
stacking_func0(data_feature, lookback_time, test_index)
}
\arguments{
\item{data_feature}{A list containing feature data for different time windows. This can be obtained from feature extraction functions such as data_feature_func_KPCA() or data_feature_func_LSTM().}

\item{lookback_time}{A vector representing the length of the cumulative time window.}

\item{test_index}{The indices of the test set, specifying which data points belong to the test set.}
}
\value{
A data frame containing the evaluation results on the test set.
}
\description{
Cum-Time Stacking Model Training and Prediction
}
