% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.CTSM.R
\name{write.config.CTSM}
\alias{write.config.CTSM}
\title{Write CTSM configuration files}
\usage{
write.config.CTSM(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
none
}
\description{
Writes config files for use with CTSM.
}
\author{
Mike Dietze, Shawn Serbin
}
