% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_params-clm.R, R/write_params_CTSM.R
\name{write_params_ctsm}
\alias{write_params_ctsm}
\title{Write CTSM Parameter File}
\usage{
write_params_ctsm(defaults = system.file("clm5_params.c171117_0001.nc",
  package = "PEcAn.CTSM"), trait.values, settings, run.id)

write_params_ctsm(defaults = system.file("clm5_params.c171117_0001.nc",
  package = "PEcAn.CTSM"), trait.values, settings, run.id)
}
\arguments{
\item{defaults}{}

\item{trait.values}{named list or data frame of traits, e.g.
\code{data.frame(vmax = 1, b0 = 2)} or \code{list(vmax = 1, b0 = 2)}}

\item{settings}{}

\item{run.id}{}

\item{defaults}{}

\item{trait.values}{named list or data frame of traits, e.g.
\code{data.frame(vmax = 1, b0 = 2)} or \code{list(vmax = 1, b0 = 2)}}

\item{settings}{}

\item{run.id}{}
}
\description{
Write CTSM Parameter File

Write CTSM Parameter File
}
